make minor changes in files

+++ src/main/java/com/example/cab/repository/CabRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CabRepository extends JpaRepository<Cab, Long> {
+public interface CabRepository extends JpaRepository<Cab, Integer> {
 
 
 }
